throttle: extract module with impl items
